requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1